# web_service/swagger.yaml
swagger: '2.0'
info:
  title: Finance-Bot (Savvify-App) API
  description: API for finance tracking, debts, analytics. Auth via Bifrost JWT.
  version: 0.6.16
basePath: /
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your token with the 'Bearer ' prefix, e.g. 'Bearer eyJhbG...'"
definitions:
  TelegramLoginRequest:
    type: object
    required:
      - telegram_data
    properties:
      telegram_data:
        type: object
        description: The JSON object returned by the Telegram Login Widget (signed).
        example:
          id: 123456789
          first_name: "TestUser"
          username: "testuser"
          auth_date: 1769679778
          hash: "Paste_Your_Hash_From_Python_Script_Here"
  AuthLoginRequest:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
        example: user@example.com
      password:
        type: string
        format: password
        example: securePassword123!
  AuthResponse:
    type: object
    properties:
      jwt:
        type: string
        description: The Bifrost Session Token (JWT)
...
paths:
...
  /login:
    post:
      tags: [Auth]
      summary: Login via Email/Password
      description: Standard login using email and password.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AuthLoginRequest'
      responses:
        '200':
          description: Login successful
          schema:
            $ref: '#/definitions/AuthResponse'
        '400':
          description: Missing email or password
        '401':
          description: Invalid credentials
  /telegram-login:
    post:
      tags: [Auth]
      summary: Login via Telegram (Widget)
      description: Exchanges signed Telegram data for a Bifrost JWT. Use generate_telegram_hash.py to create test data.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/TelegramLoginRequest'
      responses:
        '200':
          description: Login successful
          schema:
            $ref: '#/definitions/AuthResponse'
        '400':
          description: Invalid hash or missing data
        '503':
          description: Bifrost unavailable
  /verify-otp:
    post:
      tags: [Auth]
      summary: Proxy OTP verification to Bifrost
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              telegram_id:
                type: string
              code:
                type: string
      responses:
        '200':
          description: JWT token
          schema:
            type: object
            properties:
              jwt:
                type: string
  /link/complete-telegram:
    post:
      tags: [Auth]
      summary: Complete Telegram linking via token
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              telegram_id:
                type: string
              token:
                type: string
      responses:
        '200':
          description: Success message
  /link-account:
    post:
      tags: [Auth]
      summary: Link email/password credentials to current account
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        '200':
          description: Success

  # ================= TRANSACTIONS GROUP =================
...
security:
  - BearerAuth: []